+++ test/raylib/test_runner.clj
@@ -50,6 +50,7 @@
                      raylib.scenes.clockgrid-test
                      raylib.scenes.sector-test
                      raylib.scenes.palette-test
+                     raylib.scenes.gradient-test
                      raylib.easings-test
                      raylib.scenes.clock-test
                      raylib.scenes.easings-test